refactor: isolate code mode v8 runtime
