+++ codex-rs/app-server/Cargo.toml
@@ -33,6 +33,7 @@ axum = { workspace = true, default-features = false, features = [
 codex-analytics = { workspace = true }
 codex-arg0 = { workspace = true }
 codex-cloud-config = { workspace = true }
+codex-code-mode-runtime = { workspace = true }
 codex-config = { workspace = true }
 codex-core = { workspace = true }
 codex-core-plugins = { workspace = true }
