+++ codex-rs/code-mode-runtime/BUILD.bazel
@@ -1,0 +1,11 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "code-mode-runtime",
+    crate_name = "codex_code_mode_runtime",
+    crate_features = select({
+        "@rules_rs//rs/experimental/platforms/constraints:windows_msvc": [],
+        "//conditions:default": ["sandbox"],
+    }),
+    deps_extra = ["@crates//:v8"],
+)
