+++ codex-rs/code-mode-runtime/Cargo.toml
@@ -1,23 +1,29 @@
 [package]
 edition.workspace = true
 license.workspace = true
-name = "codex-code-mode"
+name = "codex-code-mode-runtime"
 version.workspace = true
 
 [lib]
 doctest = false
-name = "codex_code_mode"
+name = "codex_code_mode_runtime"
 path = "src/lib.rs"
+
+[features]
+sandbox = ["v8/v8_enable_sandbox"]
 
 [lints]
 workspace = true
 
 [dependencies]
+codex-code-mode = { workspace = true }
 codex-protocol = { workspace = true }
-serde = { workspace = true, features = ["derive"] }
+deno_core_icudata = { workspace = true }
 serde_json = { workspace = true }
-tokio = { workspace = true, features = ["sync"] }
+tokio = { workspace = true, features = ["macros", "rt", "sync", "time"] }
 tokio-util = { workspace = true, features = ["rt"] }
+tracing = { workspace = true }
+v8 = { workspace = true }
 
 [dev-dependencies]
 pretty_assertions = { workspace = true }
