+++ codex-rs/code-mode-runtime/src/runtime/globals.rs
