+++ codex-rs/core/src/session/tests.rs
@@ -4778,7 +4778,10 @@ pub(crate) async fn make_session_and_context() -> (Session, TurnContext) {
             Session::build_model_client_beta_features_header(config.as_ref()),
             /*attestation_provider*/ None,
         ),
-        code_mode_service: crate::tools::code_mode::CodeModeService::new(),
+        code_mode_session_provider: Arc::new(codex_code_mode::UnavailableCodeModeSessionProvider),
+        code_mode_service: crate::tools::code_mode::CodeModeService::new(Arc::new(
+            codex_code_mode::UnavailableCodeModeSessionProvider,
+        )),
         environment_manager: Arc::new(codex_exec_server::EnvironmentManager::default_for_tests()),
     };
 
@@ -6868,7 +6871,10 @@ where
             Session::build_model_client_beta_features_header(config.as_ref()),
             /*attestation_provider*/ None,
         ),
-        code_mode_service: crate::tools::code_mode::CodeModeService::new(),
+        code_mode_session_provider: Arc::new(codex_code_mode::UnavailableCodeModeSessionProvider),
+        code_mode_service: crate::tools::code_mode::CodeModeService::new(Arc::new(
+            codex_code_mode::UnavailableCodeModeSessionProvider,
+        )),
         environment_manager: Arc::new(codex_exec_server::EnvironmentManager::default_for_tests()),
     };
