+++ codex-rs/core/src/session/tests/guardian_tests.rs
@@ -713,6 +713,7 @@ async fn guardian_subagent_does_not_inherit_parent_exec_policy_rules() {
         plugins_manager,
         mcp_manager,
         extensions: codex_extension_api::empty_extension_registry(),
+        code_mode_session_provider: Arc::new(codex_code_mode::UnavailableCodeModeSessionProvider),
         conversation_history: InitialHistory::New,
         session_source: SessionSource::SubAgent(SubAgentSource::Other(
             GUARDIAN_REVIEWER_NAME.to_string(),
